test(slack): guard the manifest contract that broke REST end-to-end
REST didn't work end-to-end for two reasons; the backend half (the
events endpoint must answer url_verification before a signing secret
exists) is already covered. This adds the missing frontend half: the
manifest itself was structurally invalid for REST (bot_events with no
Request URL → Slack "Subscription requires a Request URL").

Extract the manifest builder into a pure slackManifest module and cover
the contract: REST declares the Events Request URL + disables Socket
Mode; Socket Mode enables the socket + omits the URL; redirect URL and
bot scopes are always present; every bot event has its matching
*:history scope; name falls back to Helix and clamps to 35 chars.

5 vitest cases pass; tsc clean.
